third commit

+++ escape.rb
@@ -1,5 +1,5 @@
 input_array = ARGV
 grav = input_array[0].to_f
-rad = input_array[1].to_i
+rad = input_array[1].to_f
 ve = Math.sqrt(2*rad*vel)
 puts ve
